feat: Separate and renamed loginNaver() to auth/openNaverSSO with using env
